Add tests helper module to reduce code duplication
