+++ tests/list_view_test.py
@@ -3,43 +3,15 @@ Test articles list template (development distribution)
 
 """
 
-import http.server
 import json
-import os
-import threading
 import unittest
-from pathlib import Path
 
-from selenium import webdriver
 from selenium.webdriver.common.by import By
 
-DIST_DIR = Path(__file__).resolve().parent.parent.joinpath("dist", "articles")
-HTTPD_PORT = 8000
+from _helpers import HTTPD_PORT, TestArticlesHelper
 
 
-class TestListView(unittest.TestCase):
-    browser: webdriver.Chrome
-
-    @staticmethod
-    def start_httpd(port: int = HTTPD_PORT):
-        os.chdir(DIST_DIR)
-        handler = http.server.SimpleHTTPRequestHandler
-        # noinspection PyTypeChecker
-        httpd = http.server.HTTPServer(("", port), handler)
-        httpd.serve_forever()
-
-    @classmethod
-    def setUpClass(cls):
-        httpd = threading.Thread(target=cls.start_httpd, daemon=True)
-        httpd.start()
-
-        options = webdriver.ChromeOptions()
-        options.add_argument("--headless")
-        cls.browser = webdriver.Chrome(options=options)
-
-    @classmethod
-    def tearDownClass(cls):
-        cls.browser.quit()
+class TestListView(TestArticlesHelper):
 
     def setUp(self):
         self.browser.get(f"http://localhost:{HTTPD_PORT}/article_list.html")
